adds  comparison sections
