+++ packages/next/cms/hooks/use-comparison.js
@@ -1,9 +1,8 @@
 import {useRouter} from "next/router";
-import {useContext, useState, useEffect} from "react";
-
+import {useState, useEffect} from "react";
+import formatProfileResponse from "../utils/formatProfileResponse";
 import axios from "axios";
 
-import ProfileContext from "../components/ProfileContext";
 
 /** */
 export default function useComparison(slug) {
@@ -11,20 +10,23 @@ export default function useComparison(slug) {
   const [comparisonLoading, setComparisonLoading] = useState(false);
   const {query, locale} = useRouter();
   const compareSlug = query?.compare;
-
   useEffect(() => {
     if (compareSlug) {
       setComparisonLoading(true);
       const url = `${process.env.NEXT_PUBLIC_CMS}profile?slug=${slug}&id=${compareSlug}&locale=${locale}`;
       axios.get(url)
-        .then(resp => {
-          setComparison(resp.data);
+        .then(formatProfileResponse)
+        .then(data => {
+          setComparison(data);
           setComparisonLoading(false);
         })
         .catch(() => {
           alert("Error loading comparison");
           setComparisonLoading(false);
         });
+    }
+    else {
+      setComparison(null);
     }
   }, [compareSlug, locale, slug]);
